requirements.txt
python-telegram-bot~=13.10
sqlalchemy~=1.4
alembic~=1.7
trueskill~=0.4
